Encoding variables
Finished encoding_keys.py, the file stores the encoding used in each variable in a dictionary so then it can be used when it is needed to make new predictions. Also finished encoding all variables keeping the order of each one. Keeping the order is important for new predictions in the future