Last Corections
Now the code works with no bugs apparently. I've organized better the code in adult.py file, splitting the code in functions and adding a main call in it. I also changed the way the interaction boxes appears, making them appear always in the center of the screen. Although the code already works as it should, I'll improve it later, simplifiying the functions and variables in order to get a cleaner code.